Create space craft class
