+++ main.rb
@@ -4,6 +4,7 @@ require_relative 'mission_reporter'
 require_relative 'mission_control'
 require_relative 'mission_plan'
 require_relative 'mission'
+require_relative 'space_craft'
 require 'pry'
 
 mission_collections = MissionControl.new(mission: Mission.new)
